Unify `PdbOwnedValue` conversions.
